requirements.txt
google-cloud-bigquery
prompt_toolkit
pygments

